requirements.txt
torch
torchvision
numpy
pillow
pyyaml
tqdm
tensorboard
